feat(Final-Exam/Find_the_Missing_Number): add algorithm to determine missing number with divisibility check and conditional output across multiple test cases

+++ Semester-1/Introduction To Programming in C Language/Week-5(2D_Array_and_Recursion)/Module-19(Problem_Solving_With_2D_Array_and_Recursion)/J_Factorial.c
@@ -8,10 +8,10 @@ long long int factorial(int n){
 }
 
 int main() {
-    int n;
-    scanf("%d", &n);
+    long long int n;
+    scanf("%lld", &n);
 
     long long int ans = factorial(n);
-    printf("%d", ans);
+    printf("%lld", ans);
     return 0;
 }
